backend/requirements.txt
fastapi>=0.109.2
uvicorn[standard]>=0.27.1
sqlalchemy>=2.0.27
asyncpg>=0.29.0
alembic>=1.13.1
redis>=5.0.1
celery>=5.3.6
boto3>=1.34.34
pydantic[email]>=2.6.1
email-validator>=2.1.0
pydantic-settings>=2.1.0
PyJWT[crypto]>=2.8.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
httpx>=0.26.0
stripe>=8.3.0
anthropic>=0.18.1
onnxruntime>=1.17.0
numpy>=1.26.4
scipy>=1.12.0
librosa>=0.10.1
soundfile>=0.12.1
pyloudnorm>=0.1.0
pytest>=8.0.2
pytest-asyncio>=0.23.5
pytest-httpx>=0.28.0
prometheus-fastapi-instrumentator>=6.1.0
structlog>=24.1.0
sentry-sdk[fastapi]>=1.40.6
cryptography>=42.0.2
slowapi>=0.1.9
c2pa-python>=0.6.0

